Complete rework of to_type
