+++ castep_outputs/parsers/md_geom_file_parser.py
@@ -1,14 +1,16 @@
 """Parse castep .md or .geom files."""
+
 from __future__ import annotations
 
 from collections import defaultdict
 from typing import TextIO, TypedDict
 
-from ..utilities.castep_res import ATOMIC_DATA_TAG, TAG_RE, get_numbers
-from ..utilities.constants import FST_D, TAG_ALIASES
-from ..utilities.datatypes import AtomIndex, ThreeByThreeMatrix, ThreeVector
-from ..utilities.filewrapper import Block
-from ..utilities.utility import add_aliases, atreg_to_index, file_or_path, log_factory, to_type
+from castep_outputs.utilities.castep_res import ATOMIC_DATA_TAG, TAG_RE, get_numbers
+from castep_outputs.utilities.constants import FST_D, TAG_ALIASES
+from castep_outputs.utilities.datatypes import AtomIndex, ThreeByThreeMatrix, ThreeVector
+from castep_outputs.utilities.filewrapper import Block
+from castep_outputs.utilities.type_conv import to_type
+from castep_outputs.utilities.utility import add_aliases, atreg_to_index, file_or_path, log_factory
 
 
 class MDAtomProps(TypedDict):
