+++ castep_outputs/tools/__init__.py
@@ -3,6 +3,5 @@
 For more advanced scripts, see ``castep_outputs_tools``.
 """
 
-
 from .get_generated_files import get_generated_files as get_generated_files
 from .md_geom_parser import MDGeomParser as MDGeomParser
