+++ test/test_binary_reader.py
@@ -8,7 +8,7 @@ import pytest
 from dump_fortran_unformatted import data_types, fake_file, raw_data, to_bytes
 
 from castep_outputs.bin_parsers.fortran_bin_parser import FortranBinaryReader
-from castep_outputs.utilities.utility import to_type
+from castep_outputs.utilities.type_conv import parse_bytes
 
 DATA_FILES = Path(__file__).parent / "data_files"
 DEFAULT_SAMPLE = (1, 2, 3, 2, 3.1, "Hello", (1.0, 3.0, 6.0))
@@ -56,7 +56,7 @@ def test_binary_file_reader(fake_file, data_types, raw_data):
     reader = FortranBinaryReader(fake_file)
 
     for datum, typ, expected in zip(reader, data_types, raw_data):
-        assert to_type(datum, typ) == expected
+        assert parse_bytes(datum, typ) == expected
 
 
 def test_actual_read():
@@ -65,15 +65,15 @@ def test_actual_read():
 
     with (DATA_FILES / "test.cst_esp").open("rb") as file:
         reader = FortranBinaryReader(file)
-        for (typ, expected) in dtypes:
+        for typ, expected in dtypes:
             assert reader.get(typ) == expected
 
         for n, datum in enumerate(reader):
             x, y = divmod(n, 16)
-            ind = to_type(datum[:8], (int, ...))
+            ind = parse_bytes(datum[:8], (int, ...))
             print(ind, (x + 1, y + 1))
             assert ind == (x + 1, y + 1)  # 1 indexed
-            res = to_type(datum[8:], (complex, ...))
+            res = parse_bytes(datum[8:], (complex, ...))
             assert len(res) == dtypes[1][1][2]
